feat: adds HasSlug trait and implements on the post model
